+++ src/createChat.ts
@@ -1,0 +1,16 @@
+import { store } from 'solid-logic'
+import { longChatPane }  from './longChatPane'
+
+
+export async function createChatThing (chatContainer, me) {
+  const created = await longChatPane.mintNew({
+    session: {
+      store
+    }
+  },
+  {
+    me,
+    newBase: chatContainer.value
+  })
+  return created.newInstance
+}
