+++ src/main.js
@@ -1,3 +1,4 @@
 export { shortChatPane } from './shortChatPane'
 export { longChatPane } from './longChatPane'
-export { getChat } from './create'
+export { getChat } from './getChat'
+
